refactor: fiscspon/directory to tsx and some other components
